updated securePath location and renamed a method
